Создан хук сортировки оферов
